kgnn-kt/kgnn_kt/utils/metrics.py
import numpy as np
from sklearn.metrics import roc_auc_score, accuracy_score

def auc(y_true, y_prob) -> float:
    y_true = np.asarray(y_true).astype(int)
    y_prob = np.asarray(y_prob).astype(float)
    if len(np.unique(y_true)) < 2:
        return float("nan")
    return float(roc_auc_score(y_true, y_prob))

def acc(y_true, y_pred) -> float:
    y_true = np.asarray(y_true).astype(int)
    y_pred = np.asarray(y_pred).astype(int)
    return float(accuracy_score(y_true, y_pred))
